pyproject.toml
[project]
name = "wakatime-bot"
version = "0.0.1"
description = "Telegram bot for WakaTime API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.13.1",
    "aiohttp>=3.10.9",
    "aiosqlite>=0.20.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.6.9",
    "sqlalchemy>=2.0.35",
]
